Simple base with Enemy NavMesh pathfinding
I have made a "simple base" which consists of 4 walls. I made an Enemy prefab which uses NavMesh to get to the targeted destination, just drag the empty game object with a transform to the 'Move Position' in the inspector of the Enemy prefab to get the targeted position. I also cleaned up my scene after scraping my track level, you can try out the stuff in the scene named 'Martins Passthrough'